Added a lot of initial functionality for Admin Page | Overhaul
Added functionality for adding to or deleting from sites, and adding or deleting sites for admins from within the app. This is more or less a large basic framework for future improvement and will be tested asap when firebase rules are changed to work with it.

Admins can change existing site data and images and can create entirely new sites with full descriptions and images!